start migrating length/capacity over to long

+++ src/main/java/uk/co/real_logic/agrona/concurrent/MappedResizeableBuffer.java
@@ -15,28 +15,18 @@
  */
 package uk.co.real_logic.agrona.concurrent;
 
-import static uk.co.real_logic.agrona.BitUtil.SIZE_OF_BYTE;
-import static uk.co.real_logic.agrona.BitUtil.SIZE_OF_CHAR;
-import static uk.co.real_logic.agrona.BitUtil.SIZE_OF_DOUBLE;
-import static uk.co.real_logic.agrona.BitUtil.SIZE_OF_FLOAT;
-import static uk.co.real_logic.agrona.BitUtil.SIZE_OF_INT;
-import static uk.co.real_logic.agrona.BitUtil.SIZE_OF_LONG;
-import static uk.co.real_logic.agrona.BitUtil.SIZE_OF_SHORT;
-import static uk.co.real_logic.agrona.UnsafeAccess.UNSAFE;
-import static uk.co.real_logic.agrona.concurrent.UnsafeBuffer.ALIGNMENT;
-import static uk.co.real_logic.agrona.concurrent.UnsafeBuffer.ARRAY_BASE_OFFSET;
-import static uk.co.real_logic.agrona.concurrent.UnsafeBuffer.NATIVE_BYTE_ORDER;
-import static uk.co.real_logic.agrona.concurrent.UnsafeBuffer.NULL_BYTES;
-import static uk.co.real_logic.agrona.concurrent.UnsafeBuffer.SHOULD_BOUNDS_CHECK;
+import uk.co.real_logic.agrona.DirectBuffer;
+import uk.co.real_logic.agrona.IoUtil;
+import uk.co.real_logic.agrona.MutableDirectBuffer;
 
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
 import java.nio.channels.FileChannel;
 import java.nio.charset.StandardCharsets;
 
-import uk.co.real_logic.agrona.DirectBuffer;
-import uk.co.real_logic.agrona.IoUtil;
-import uk.co.real_logic.agrona.MutableDirectBuffer;
+import static uk.co.real_logic.agrona.BitUtil.*;
+import static uk.co.real_logic.agrona.UnsafeAccess.UNSAFE;
+import static uk.co.real_logic.agrona.concurrent.UnsafeBuffer.*;
 
 /**
  * Supports regular, byte ordered, and atomic (memory ordered) access to an underlying buffer.
@@ -175,9 +165,9 @@ public class MappedResizeableBuffer implements AtomicBuffer, AutoCloseable
         UNSAFE.setMemory(null, addressOffset + index, length, value);
     }
 
-    public int capacity()
+    public long capacity()
     {
-        return (int) capacity;
+        return capacity;
     }
 
     public void checkLimit(final long limit)
